% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVViewer.R
\docType{package}
\name{CSVViewer-package}
\alias{CSVViewer}
\alias{CSVViewer-package}
\title{CSVViewer: Shiny App to view csv files}
\description{
Shiny App to review csv files and add comments on these files. Documentation for each of the csv files can be added and viewed in the app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EHDEN/CSVViewer}
  \item Report bugs at \url{https://github.com/EHDEN/CSVViewer/issues}
}

}
\author{
\strong{Maintainer}: Peter Rijnbeek \email{p.rijnbeek@ehden.eu}

}
\keyword{internal}
