% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{launch}
\alias{launch}
\title{Launch the application}
\usage{
launch(app_port = 8888)
}
\arguments{
\item{app_port}{the port on which the app is available}
}
\value{
nothing
}
\description{
Launch the application
}
