% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Launch.R
\name{launch_demo}
\alias{launch_demo}
\title{Launch the application}
\usage{
launch_demo(app_port = 8888, csvfolder = tempdir())
}
\arguments{
\item{app_port}{the port on which the app is available}

\item{csvfolder}{to which the example files are copied and the app is initated with}
}
\value{
nothing
}
\description{
Launch the application
}
